% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/comments.R
\name{add_comment}
\alias{add_comment}
\title{comment on a Google Drive file}
\usage{
add_comment(token, file_id, comment_text, ...)
}
\arguments{
\item{token}{a token, generated with \code{\link{driver_connect}}.}

\item{file_id}{the ID of a file - see \code{\link{file_metadata}} for further discussion.}

\item{comment_text}{a string containing the comment you wish to leave}

\item{...}{further arguments to pass to httr's POST}
}
\value{
a basic comment object, similar to those returned from \code{\link{get_comment}}
}
\description{
add a comment to a Google Drive file. These comments are not (currently) anchored,
meaning that they're associated with the file as a whole rather than tied to any particular element
or line of text.
}
\examples{
\dontrun{
file_id <- list_files(token, max_results = 1)$items[[1]]$id
comment_metadata <- add_comment(token, file_id, comment = text = "I have strong negative opinions
                               about this proposal but no actual idea for how to improve it")
}
}
\seealso{
\code{\link{upload_file}} for uploading a file, and \code{\link{add_reply}} for replying
to an existing comment.
}

