% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/permission.r
\name{add_permission}
\alias{add_permission}
\title{Add permission for a file}
\usage{
add_permission(file_id, email, role = "reader", type = "user",
  token = get_access_cred(), ...)
}
\arguments{
\item{file_id}{the ID of a file}

\item{email}{email address to be added}

\item{role}{the primary role of the user}

\item{type}{the account type}

\item{token}{a valid OAuth2.0 token}

\item{...}{any other argument to be passed to the API call}
}
\value{
permission id
}
\description{
Add permission for a file
}

