% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/replies.R
\name{add_reply}
\alias{add_reply}
\title{Reply to a comment}
\usage{
add_reply(file_id, comment_id, reply_text, ...)
}
\arguments{
\item{file_id}{the ID of a file - or the full URL for accessing it via your browser.
See \code{\link{file_metadata}} for further discussion.}

\item{comment_id}{the ID of a comment, which can be found in the output of \code{\link{list_comments}}}

\item{reply_text}{a string containing the reply you wish to add}

\item{...}{further arguments to pass to httr's POST}
}
\value{
a basic reply object, similar to those returned from \code{\link{get_reply}}
}
\description{
replies to a specific comment on a Google Drive file
}
\seealso{
\code{\link{upload_file}} for uploading a file, and \code{\link{add_comment}} for adding
a new comment
}

