% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/azuga.R
\name{azuga_ftp_list}
\alias{azuga_ftp_list}
\title{Azuga FTP List}
\usage{
azuga_ftp_list(
  path = "",
  host = "ftp://pentaho8.azuga.com",
  user = Sys.getenv("AZUGA_FTP_USER"),
  pass = Sys.getenv("AZUGA_FTP_PASS")
)
}
\arguments{
\item{path}{directory to list}

\item{host}{domain name of Azuga FTP server}

\item{user}{username}

\item{pass}{password}
}
\value{
tibble representation of the list results
}
\description{
Azuga FTP List
}
\examples{
\dontrun{
azuga_ftp_list("Reports")
}
}
