% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/azuga.R
\name{azuga_ftp_parse_date}
\alias{azuga_ftp_parse_date}
\title{Azuga FTP Parse Date}
\usage{
azuga_ftp_parse_date(x)
}
\arguments{
\item{x}{string representation of the datetime from the ftp server}
}
\value{
datetime representation of the datetime from the ftp server
}
\description{
Azuga FTP Parse Date
}
\examples{
# Two versions of a datetime observed from the ftp server

## Version 1
azuga_ftp_parse_date("Apr 04 2021")

## Version 2
azuga_ftp_parse_date("Apr 04 07:39")
}
