% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unsorted.R
\name{bdiag}
\alias{bdiag}
\title{Create Block Diagonal Matrix from Passed Matricies}
\usage{
bdiag(...)
}
\arguments{
\item{...}{matricies to create block diagonal matrix from}
}
\value{
matrix
}
\description{
Copied from similarly named function in dlm package
}
\examples{
x <- matrix(runif(10), 5, 2)
y <- matrix(runif(6), 3, 2)
bdiag(x,y)
}
