% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app.R, R/census.R
\name{app_census_block_group_population}
\alias{app_census_block_group_population}
\alias{census_block_group_population}
\title{Shiny App to Lookup Urban and Rural Data for a Given Census Block Group}
\usage{
app_census_block_group_population()

census_block_group_population(
  fips_code_12,
  vintage = c("2010", "2000"),
  key = Sys.getenv("CENSUS_KEY")
)
}
\arguments{
\item{fips_code_12}{12-digit FIPS code}

\item{vintage}{Year of dataset, e.g., 2000. This data is only collected
during decennial censuses.}

\item{key}{Your Census API key, requested using
\url{https://api.census.gov/data/key_signup.html}.}
}
\value{
A tibble with FIPS codes and P2 variables for the given FIPS code.
}
\description{
Shiny App to Lookup Urban and Rural Data for a Given Census Block Group

Lookup Urban and Rural Data for a Given Census Block Group
}
\examples{
\dontrun{
## each new state in a session requires a download from the Census Bureau
census_block_group_population("295101022001")
census_block_group_population("171635033341")

## looking up FIPS codes for another vintage
census_block_group_population("295101022001", vintage = "2000")
}
}
