% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app.R, R/census.R
\name{app_census_geocode_components}
\alias{app_census_geocode_components}
\alias{census_geocode_components}
\title{Shiny App to Geocode a single address}
\usage{
app_census_geocode_components()

census_geocode_components(
  street,
  city,
  state,
  postal_code,
  vintage = c("2010", "2000"),
  key = Sys.getenv("GEOCODIO_API_KEY")
)
}
\arguments{
\item{street, city, state, postal_code}{address components}

\item{vintage}{year of dataset, e.g., 2000.}

\item{key}{\href{https://geocod.io/}{geocodio} api key}
}
\value{
a \code{\link[tibble:tibble]{tibble::tibble()}} containing geocodio data per the request
}
\description{
Shiny App to Geocode a single address

Geocode a single address
}
\seealso{
\code{\link[rgeocodio:gio_geocode_components]{rgeocodio::gio_geocode_components()}},
\href{https://geocod.io/docs/#fields}{geocodio documentation} for more information
on costs/pricing.
}
