% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app.R, R/census.R
\name{app_census_zipcode_population}
\alias{app_census_zipcode_population}
\alias{census_zipcode_population}
\title{Shiny App to Lookup Urban and Rural Data for a Given 5-Digit Zip Code}
\usage{
app_census_zipcode_population()

census_zipcode_population(
  zip_code_5,
  vintage = c("2010", "2000"),
  key = Sys.getenv("CENSUS_KEY")
)
}
\arguments{
\item{zip_code_5}{five-digit zip code}

\item{vintage}{Year of dataset, e.g., 2000. This data is only collected
during decennial censuses.}

\item{key}{Your Census API key, requested using
\url{https://api.census.gov/data/key_signup.html}.}
}
\value{
A tibble with population data for a zip code tabulation area.
}
\description{
Shiny App to Lookup Urban and Rural Data for a Given 5-Digit Zip Code

Lookup Urban and Rural Data for a Given 5-Digit Zip Code
}
\examples{
\dontrun{
## only one download from Census per session per vintage
census_zipcode_population("63130")
census_zipcode_population("63102")

## another vintage
census_zipcode_population("63130", vintage = "2000")
}
}
