% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/census.R
\name{census_zipcode_populations}
\alias{census_zipcode_populations}
\title{Retrieve Urban and Rural (p2) Summary File 1 (sf1) Census data}
\usage{
census_zipcode_populations(
  vintage = c("2010", "2000"),
  key = Sys.getenv("CENSUS_KEY")
)
}
\arguments{
\item{vintage}{= Year of dataset, e.g., 2000. This data is only collected
during decennial censuses.}

\item{key}{= Your Census API key, requested using
\url{https://api.census.gov/data/key_signup.html}.}
}
\value{
A tibble with P2 variables by zip code tabulation area.
}
\description{
Retrieve Urban and Rural (p2) Summary File 1 (sf1) Census data
}
\details{
Decennial Census API Documentation
\url{https://www.census.gov/data/developers/data-sets/decennial-census.html}

URBAN AND RURAL variables listed in:
\url{https://api.census.gov/data/2010/dec/sf1/variables.html}
\tabular{llllllll}{
\strong{Name} \tab
\strong{Label} \tab
\strong{Concept} \tab
\strong{Required} \tab
\strong{Attributes} \tab
\strong{Limit} \tab
\strong{Predicate Type} \tab
\strong{Group} \cr
P002001 \tab
Total \tab
URBAN AND RURAL \tab
not required \tab
P002001ERR \tab
0 \tab
int  \tab
P2 \cr
P002002 \tab
Total!!Urban \tab
URBAN AND RURAL \tab
not required \tab
\tab
0 \tab
int  \tab
P2 \cr
P002003 \tab
Total!!Urban!!Inside urbanized areas \tab
URBAN AND RURAL \tab
not required \tab
\tab
0 \tab
int  \tab
P2 \cr
P002004 \tab
Total!!Urban!!Inside urban clusters \tab
URBAN AND RURAL \tab
not required \tab
\tab
0 \tab
int  \tab
P2 \cr
P002005 \tab
Total!!Rural \tab
URBAN AND RURAL \tab
not required \tab
\tab
0 \tab
int  \tab
P2
}
}
\examples{
\dontrun{
## data for all zip code tabulation areas
census_zipcode_populations()

## compute percents urban and rural for each zip code tabulation area
census_zipcode_populations() \%>\%
  mutate(
    cen_zcta_pct_urban_2010 =
      100 * cen_zcta_pop_urban_2010 / cen_zcta_pop_total_2010,
    cen_zcta_pct_rural_2010 =
      100 * cen_zcta_pop_rural_2010 / cen_zcta_pop_total_2010
  )
}

}
