% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unsorted.R
\name{center}
\alias{center}
\title{Center matrix (subtract row mean from each row)}
\usage{
center(x)
}
\arguments{
\item{x}{vector or matrix}
}
\value{
matrix with centered rows
}
\description{
After centering, rowsums sum to zero. Particularly useful when working
with categorical data in log-space.
}
\examples{
x <- matrix(rnorm(100), 10, 10)
y <- center(x)
rowSums(y)
}
