% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/breadcrumbs.R
\name{collate_daily_breadcrumb_metadata}
\alias{collate_daily_breadcrumb_metadata}
\title{Collate Daily Breadcrumb Metadata}
\usage{
collate_daily_breadcrumb_metadata(breadcrumb_directory)
}
\arguments{
\item{breadcrumb_directory}{directory containing the breadcrumb files}
}
\value{
tibble containing \code{file} as full path to each breadcrumb and \code{date}
as extracted from the breadcrumb file name.
}
\description{
Collate Daily Breadcrumb Metadata
}
\examples{
\dontrun{
ris_storage_path("Reports") \%>\%
  collate_daily_breadcrumb_metadata()
}
}
