% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/breadcrumbs.R
\name{collate_monthly_breadcrumbs}
\alias{collate_monthly_breadcrumbs}
\title{Collate Monthly Breadcrumbs}
\usage{
collate_monthly_breadcrumbs(monthly_breadcrumb_files, debug = FALSE)
}
\arguments{
\item{monthly_breadcrumb_files}{names of the monthly breadcrumb files to
collate (i.e., return value from \code{write_monthly_breadcrumbs_for_year()})}
}
\description{
Collate Monthly Breadcrumbs
}
\examples{
\dontrun{
ris_storage_path("Reports") \%>\%
  collate_daily_breadcrumb_metadata() \%>\%
  write_monthly_breadcrumbs_for_year(2015) \%>\%
  collate_monthly_breadcrumbs()
}
}
