% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crumbs.R
\name{collate_monthly_crumbs}
\alias{collate_monthly_crumbs}
\title{Collate Monthly Breadcrumbs}
\usage{
collate_monthly_crumbs(monthly_crumb_files, debug = FALSE)
}
\arguments{
\item{monthly_crumb_files}{names of the monthly breadcrumb files to
collate (i.e., return value from \code{write_monthly_crumbs_for_year()})}

\item{debug}{output word counts for all monthly and the year breadcrumb files
to see if the montly add up to the yearly}
}
\description{
Collate Monthly Breadcrumbs
}
\examples{
\dontrun{
ris_storage_path("Reports") \%>\%
  collate_daily_crumb_metadata() \%>\%
  write_monthly_crumbs_for_year(2015) \%>\%
  collate_monthly_crumbs()
}
}
