% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/array_munging.R
\name{any_cases_array}
\alias{any_cases_array}
\alias{all_cases_array}
\title{Test for any/all non-NA along margin}
\usage{
any_cases_array(a, margin = 1)

all_cases_array(a, margin = 1)
}
\arguments{
\item{a}{multidimensional array}

\item{margin}{margins to test along}
}
\value{
matrix
}
\description{
\code{complete_cases_array} is generalization of \code{complete_cases}
to arrays. \code{any_cases_array} tests for any non-NA rather than all
non-NA along margin.
}
\examples{
a <- array(0, dim =c(3,4,2))
a[2,2,] <- NA
any_cases_array(a, margin=1)
any_cases_array(a, margin=c(1,2))
all_cases_array(a, margin=1)
all_cases_array(a, margin=c(1,2))
}
