% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/complex_coda.R
\name{convert_coda_covariance}
\alias{convert_coda_covariance}
\alias{ilrvar2phi}
\alias{ilrvar2ilrvar}
\alias{ilrvar2clrvar}
\alias{clrvar2ilrvar}
\alias{clrvar2varmat}
\alias{ilrvar2varmat}
\alias{alrvar2clrvar}
\alias{clrvar2alrvar}
\alias{alrvar2alrvar}
\alias{alrvar2ilrvar}
\alias{ilrvar2alrvar}
\alias{alrvar2varmat}
\title{Convert CoDA covariance matricies between representations}
\usage{
ilrvar2phi(Sigma, V)

ilrvar2ilrvar(Sigma, V1, V2)

ilrvar2clrvar(Sigma, V)

clrvar2ilrvar(Sigma, V)

clrvar2varmat(Sigma)

ilrvar2varmat(Sigma, V)

alrvar2clrvar(Sigma, d1)

clrvar2alrvar(Sigma, d2)

alrvar2alrvar(Sigma, d1, d2)

alrvar2ilrvar(Sigma, d1, V2)

ilrvar2alrvar(Sigma, V1, d2)

alrvar2varmat(Sigma, d1)
}
\arguments{
\item{Sigma}{covariance matrix in specified transformed space}

\item{V}{ILR contrast matrix (i.e., transformation matrix of ILR)}

\item{V1}{ILR contrast matrix of basis Sigma is already in}

\item{V2}{ILR contrast matrix of basis Sigma is desired in}

\item{d1}{alr reference element Sigma is already expressed with respec to}

\item{d2}{alr reference element Sigma is to be expressed with respect to}
}
\value{
matrix
}
\description{
\code{ilrvar}, \code{clrvar}, and \code{varmat} (variation matrix).
\code{ilrvar2phi} calculates phi statistics (for proportionality)
from an ILR covariance matrix as described in Lovell (2015).
}
\examples{
x <- matrix(runif(30),  10, 3)
x <- miniclo(x)
x.ilr <- ilr(x)
V <- create_default_ilr_base(3)
Sigma <- cov(x.ilr)
Sigma.clr <- ilrvar2clrvar(Sigma, V)
clrvar2ilrvar(Sigma.clr, V)
clrvar2varmat(Sigma.clr)
ilrvar2varmat(Sigma, V)
ilrvar2phi(Sigma,V)
}
