% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/files.R
\name{copy_file}
\alias{copy_file}
\title{copy a Google Drive file}
\usage{
copy_file(token, file_id, ...)
}
\arguments{
\item{token}{a token, generated with \code{\link{driver_connect}}.}

\item{file_id}{the ID of a file - or the full URL for accessing it via your browser.
See \code{\link{file_metadata}} for further discussion.}

\item{...}{further arguments to pass to httr's POST.}
}
\value{
a set of metadata associated with the copy of the file, matching
the output of \code{\link{file_metadata}}.
}
\description{
takes a Google Drive file and creates a copy of it, with the same
access restrictions.
}

