% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/breadcrumbs.R
\name{create_empty_breadcrumb_database_table}
\alias{create_empty_breadcrumb_database_table}
\title{Create Empty Breadcrumb Database Table}
\usage{
create_empty_breadcrumb_database_table(con, year, fidelity = c("30", "1"))
}
\arguments{
\item{con}{a DBIConnection object, as returned by \code{dbConnect()}.}

\item{year}{year of the table}

\item{fidelity}{fidelty of the breadcrumbs
(i.e., either 30 seconds or 1 second)}
}
\value{
name of the newly created database table
}
\description{
Create Empty Breadcrumb Database Table
}
\examples{
\dontrun{
con <- DBI::dbConnect(
  RMariaDB::MariaDB(),
  default.file = normalizePath("~./.my.cnf"),
  group = "spark.myftp.org"
)

create_empty_breadcrumb_database_table(con, 2015)

DBI::dbDisconnect(con)
}
}
