% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core_functions.R
\name{create_features_df}
\alias{create_features_df}
\title{Create Data Frame of Features for Driverness Scoring}
\usage{
create_features_df(
  annovar_csv_path,
  scna_df,
  phenolyzer_annotated_gene_list_path,
  prep_phenolyzer_input = FALSE,
  log2_ratio_threshold = 0.25,
  gene_overlap_threshold = 25,
  MCR_overlap_threshold = 25,
  hotspot_threshold = 5L,
  log2_hom_loss_threshold = -1
)
}
\arguments{
\item{annovar_csv_path}{path to 'ANNOVAR' csv output file}

\item{scna_df}{the SCNA segments data frame. Must contain: \describe{
\item{chr}{chromosome the segment is located in}
\item{start}{start position of the segment}
\item{end}{end position of the segment}
\item{log2ratio}{\ifelse{html}{\out{log<sub>2</sub>}}{\eqn{log_2}} ratio of the segment}
}}

\item{phenolyzer_annotated_gene_list_path}{path to 'phenolyzer'
"annotated_gene_list" file}

\item{prep_phenolyzer_input}{boolean to indicate whether or not to create
a vector of genes for use as the input of 'phenolyzer' (default = \code{FALSE}).
If \code{TRUE}, the features data frame is not created and instead the vector
of gene symbols (union of all genes for which scores are available) is
returned.}

\item{log2_ratio_threshold}{the \ifelse{html}{\out{log<sub>2</sub>}}{\eqn{log_2}}
ratio threshold for keeping high-confidence SCNA events (default = 0.25)}

\item{gene_overlap_threshold}{the percentage threshold for the overlap between
a segment and a transcript (default = 25). This means that if only a segment
overlaps a transcript more than this threshold, the transcript is assigned
the segment's SCNA event.}

\item{MCR_overlap_threshold}{the percentage threshold for the overlap between
a gene and an MCR region (default = 25). This means that if only a gene
overlaps an MCR region more than this threshold, the gene is assigned the
MCR's SCNA density}

\item{hotspot_threshold}{to determine hotspot genes, the (integer) threshold
for the minimum number of cases with certain mutation in COSMIC (default = 5)}

\item{log2_hom_loss_threshold}{to determine double-hit events, the
\ifelse{html}{\out{log<sub>2</sub>}}{\eqn{log_2}} threshold for identifying
homozygous loss events (default = -1).}
}
\value{
If \code{prep_phenolyzer_input=FALSE} (default), a data frame of
features for scoring cancer driverness of each gene (\code{gene_symbol} as
the first column and 26 other columns containing features). If
\code{prep_phenolyzer_input=TRUE}, the functions returns a vector gene symbols
(union of all gene symbols for which scores are available) to be used as the
input for performing 'phenolyzer' analysis.
}
\description{
Create Data Frame of Features for Driverness Scoring
}
\examples{
\dontrun{
path2annovar_csv <- system.file("extdata/example.hg19_multianno.csv",
                                package = "driveR")
path2phenolyzer_out <- system.file("extdata/example.annotated_gene_list",
                                   package = "driveR")
features_df <- create_features_df(annovar_csv_path = path2annovar_csv,
                                  scna_df = example_scna_table,
                                  phenolyzer_annotated_gene_list_path = path2phenolyzer_out)
}
}
