% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/files.R
\name{create_file}
\alias{create_file}
\title{create an empty file, or upload a local file}
\usage{
create_file(token, file_path = NULL, ...)
}
\description{
create an empty file, or upload a local file
}

