% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{create_SCNA_score_df}
\alias{create_SCNA_score_df}
\title{Create SCNA Score Data Frame}
\usage{
create_SCNA_score_df(
  scna_df,
  log2_ratio_threshold = 0.25,
  gene_overlap_threshold = 25,
  MCR_overlap_threshold = 25
)
}
\arguments{
\item{scna_df}{the SCNA segments data frame. Must contain: \describe{
\item{chr}{chromosome the segment is located in}
\item{start}{start position of the segment}
\item{end}{end position of the segment}
\item{log2ratio}{\ifelse{html}{\out{log<sub>2</sub>}}{\eqn{log_2}} ratio of the segment}
}}

\item{log2_ratio_threshold}{the \ifelse{html}{\out{log<sub>2</sub>}}{\eqn{log_2}}
ratio threshold for keeping high-confidence SCNA events (default = 0.25)}

\item{gene_overlap_threshold}{the percentage threshold for the overlap between
a segment and a transcript (default = 25). This means that if only a segment
overlaps a transcript more than this threshold, the transcript is assigned
the segment's SCNA event.}

\item{MCR_overlap_threshold}{the percentage threshold for the overlap between
a gene and an MCR region (default = 25). This means that if only a gene
overlaps an MCR region more than this threshold, the gene is assigned the
MCR's SCNA density}
}
\value{
data frame containing gene-level SCNA events (that are concordant with
MCRs) with proxy SCNA scores (\code{CNVdensity..CNV.Mb.})

data frame of SCNA proxy scores containing 2 columns: \describe{
\item{gene_symbol}{HGNC gene symbol}
\item{SCNA_density}{SCNA proxy score. SCNA density (SCNA/Mb) of the minimal common region (MCR) in which the gene is located.}
}
}
\description{
Create SCNA Score Data Frame
}
\details{
The function first aggregates SCNA \ifelse{html}{\out{log<sub>2</sub>}}{\eqn{log_2}} ratio
on gene-level (by keeping the ratio with the maximal \ifelse{html}{\out{|log<sub>2</sub>|}}{\eqn{|log_2|}}
ratio over all the SCNA segments overlapping a gene). Next, it identifies the
minimal common regions (MCRs) that the genes overlap and finally assigns the
SCNA density (SCNA/Mb) values as proxy SCNA scores.
}
\examples{
\dontrun{
SCNA_scores_df <- driveR:::create_SCNA_score_df(imielinski_scna_table)
}
}
