% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/breadcrumbs.R
\docType{data}
\name{daily_breadcrumb_headers}
\alias{daily_breadcrumb_headers}
\title{Daily Breadcrumb Headers}
\format{
An object of class \code{character} of length 7.
}
\usage{
daily_breadcrumb_headers
}
\description{
Breadcrumb files are inconsistently formatted. Known issues include that
some are comma delimited and some semicolon, some contain datetimes and some
contain dates and times, some use '' and some use '.' as missing, and some
included fields are inconsistent from one file to the next - especially the
datetimes vs date and time issue. This character vector contains all known
breadcrumb file headers and is useful in determining how to read a daily
breadcrumb file in \code{read_daily_breadcrumb()}.
}
\keyword{datasets}
