% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crumbs.R
\docType{data}
\name{daily_crumb_col_types}
\alias{daily_crumb_col_types}
\title{Daily Breadcrumb Column Types}
\format{
An object of class \code{list} of length 7.
}
\usage{
daily_crumb_col_types
}
\description{
List of functions that each provide the col spec for the given header class.
}
\keyword{datasets}
