% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crumbs.R
\docType{data}
\name{daily_crumb_name_patterns}
\alias{daily_crumb_name_patterns}
\title{Daily Breadcrumb Name Patterns}
\format{
An object of class \code{character} of length 4.
}
\usage{
daily_crumb_name_patterns
}
\description{
Breadcrumb files are inconsistently named. This character vector contains
regular expressions for all known name patterns. This is useful for
determining how to extract the date from the file name in
\code{collate_all_daily_crumb_metadata()}.
}
\keyword{datasets}
