% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crumbs.R
\docType{data}
\name{daily_crumb_readers}
\alias{daily_crumb_readers}
\title{Daily Breadcrumb Readers}
\format{
An object of class \code{list} of length 7.
}
\usage{
daily_crumb_readers
}
\description{
List of functions that each can read the desired breadcrumb type.
}
\keyword{datasets}
