% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/comments.R
\name{delete_comment}
\alias{delete_comment}
\title{delete a comment on a Google Drive file}
\usage{
delete_comment(token, file_id, comment_id, ...)
}
\arguments{
\item{token}{a token, generated with \code{\link{driver_connect}}.}

\item{file_id}{the ID of a file - see \code{\link{file_metadata}} for further discussion.}

\item{comment_id}{the ID of a comment, which can be easily retrieved with \code{\link{list_comments}}}

\item{...}{further arguments to pass to httr's DELETE}
}
\value{
TRUE if the comment was successfully deleted, FALSE otherwise.
}
\description{
\code{delete_comment}, when provided with an authentication token, a file ID
and a comment ID, will let you delete a comment automatically. It's worth noting that this will
/also/ junk any replies to the initial comment.
}
\examples{
\dontrun{
#Grab a list of comments in a file and delete the first one
comments <- list_comments(token = token, file_id = "f38rggruFKJad30")
delete_comment(token, "f38rggruFKJad30", comments$items[[1]]$id)
}
}
\seealso{
\code{\link{list_comments}} to retrieve comment metadata (including the comment
IDs) for a file.
}

