% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/revisions.R
\name{delete_revision}
\alias{delete_revision}
\title{remove a particular revision of a Google Drive file}
\usage{
delete_revision(file_id, rev_id, ...)
}
\arguments{
\item{file_id}{the ID of a file - or the full URL for accessing it via your browser.
See \code{\link{file_metadata}} for further discussion.}

\item{rev_id}{the ID of a revision of that file.}

\item{...}{further arguments to pass to httr's DELETE.}
}
\description{
junks a specified revision of a Google Drive file. Note that some file types
may not support revision deletion, in which case a 400 error will be returned.
}

