% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{determine_double_hits}
\alias{determine_double_hits}
\title{Determine Double-Hit Genes}
\usage{
determine_double_hits(
  annovar_csv_path,
  scna_df,
  gene_overlap_threshold = 25,
  log2_threshold = -1
)
}
\arguments{
\item{annovar_csv_path}{path to ANNOVAR csv output file}

\item{scna_df}{the SCNA segments data frame. Must contain: \describe{
\item{chr}{chromosome the segment is located in}
\item{start}{start position of the segment}
\item{end}{end position of the segment}
\item{log2ratio}{\ifelse{html}{\out{log<sub>2</sub>}}{\eqn{log_2}} ratio of the segment}
}}

\item{gene_overlap_threshold}{the percentage threshold for the overlap between
a segment and a transcript (default = 25). This means that if only a segment
overlaps a transcript more than this threshold, the transcript is assigned
the segment's SCNA event.}

\item{log2_threshold}{\ifelse{html}{\out{log<sub>2</sub>}}{\eqn{log_2}} threshold
for determining homozygous loss events (default = -1)}
}
\value{
vector of gene symbols that are subject to double-hit event(s), i.e.
non-synonymous mutation + homozygous CN loss
}
\description{
Determine Double-Hit Genes
}
\examples{
path2annovar_csv <- system.file("extdata/imielinski.hg19_multianno.csv",
                                package = "driveR")
\dontrun{
dhit_genes <- driveR:::determine_double_hits(path2annovar_csv, imielinski_scna_table)
}
}
