% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{determine_hotspot_genes}
\alias{determine_hotspot_genes}
\title{Determine Hotspot Containing Genes}
\usage{
determine_hotspot_genes(annovar_csv_path, threshold = 5L)
}
\arguments{
\item{annovar_csv_path}{path to ANNOVAR csv output file}

\item{threshold}{(integer) threshold for the minimum number of cases with
a certain mutation in COSMIC (default = 5)}
}
\value{
vector of gene symbols of genes containing hotspot mutation(s)
}
\description{
Determine Hotspot Containing Genes
}
\examples{
path2annovar_csv <- system.file("extdata/imielinski.hg19_multianno.csv",
                                package = "driveR")
hotspot_genes <- driveR:::determine_hotspot_genes(path2annovar_csv)
}
