% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simple_statistics.R
\name{dmode}
\alias{dmode}
\title{Mode of a vector x}
\usage{
dmode(x)
}
\arguments{
\item{x}{vector of values}
}
\value{
vector (can return multiple maxima)
}
\description{
Mode of a vector x
}
\examples{
dmode(c(2,3,34,2,2))
}
