% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/authentication.R
\name{driver_connect}
\alias{driver_connect}
\title{connect to Google Drive and authorise driver to access your files.}
\usage{
driver_connect()
}
\description{
\code{driver_connect} produces an OAuth 2.0 token which can be passed
into other functions in the driver package, enabling driver to access your Google Drive
files on your behalf.

#'
}
\examples{
\dontrun{
#Use driver_connect to generate a token which is than quietly saved and called internally
in other driver functions.
driver_connect()
}
}

