% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/http_wrappers.R
\name{driver_delete}
\alias{driver_delete}
\title{send a DELETE request to the Google Drive API}
\usage{
driver_delete(parameters, token, ...)
}
\arguments{
\item{parameters}{a constructed query string passed from one of the exported functions.}

\item{token}{a token retrieved with \code{\link{driver_connect}}}

\item{...}{further arguments to httr's DELETE function.}
}
\description{
send a DELETE request to the Google Drive API
}

