% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/http_wrappers.R
\name{driver_get}
\alias{driver_get}
\title{send a GET request to the Google Drive API}
\usage{
driver_get(parameters, out_class, ...)
}
\arguments{
\item{parameters}{a constructed query string passed from one of the exported functions.}

\item{out_class}{a classname to set, used to distinguish between simplifier methods}

\item{...}{further arguments to httr's GET function.}
}
\description{
send a GET request to the Google Drive API
}

