% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/http_wrappers.R
\name{driver_patch}
\alias{driver_patch}
\title{send a PATCH request to the Google Drive API}
\usage{
driver_patch(parameters, token, ...)
}
\arguments{
\item{parameters}{a constructed query string passed from one of the exported functions.}

\item{token}{a token retrieved with \code{\link{driver_connect}}}

\item{...}{further arguments to httr's PATCH function.}
}
\description{
send a PATCH request to the Google Drive API
}

