% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/http_wrappers.R
\name{driver_post}
\alias{driver_post}
\title{send a POST request to the Google Drive API}
\usage{
driver_post(parameters, ...)
}
\arguments{
\item{parameters}{a constructed query string passed from one of the exported functions.}

\item{...}{further arguments to httr's POST function.}
}
\description{
send a POST request to the Google Drive API
}

