% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/http_wrappers.R
\name{driver_put}
\alias{driver_put}
\title{send a PUT request to the Google Drive API}
\usage{
driver_put(parameters, ...)
}
\arguments{
\item{parameters}{a constructed query string passed from one of the exported functions.}

\item{...}{further arguments to httr's PUT function.}
}
\description{
send a PUT request to the Google Drive API
}

