% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/files.R
\name{empty_trash}
\alias{empty_trash}
\title{Empties the trash}
\usage{
empty_trash(...)
}
\arguments{
\item{...}{arguments to pass to httr's DELETE.}
}
\value{
TRUE if the trash was successfully emptied, FALSE otherwise.
}
\description{
empties the user's Google Drive "trash" folder.
}

