% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coda_units.R
\name{eu2logfoldchange}
\alias{eu2logfoldchange}
\title{Calculate Equilvalent Log Fold Change for Given Change in Evidence Information}
\usage{
eu2logfoldchange(x, r = 1, s = 1)
}
\arguments{
\item{x}{Evidence Information Units}

\item{r}{Number of parts in numerator of log-ratio}

\item{s}{number of parts in denominator of log-ratio}
}
\value{
scalar
}
\description{
Calculate Equilvalent Log Fold Change for Given Change in Evidence Information
}
\details{
EI is defined on log-ratio scale and for a given balance can be
derived as \deqn{y = sqrt(rs/(r+s))log(Gm(X^+)/Gm(X^-))}
}
\examples{
eu2logfoldchange(1)
}
