% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{example_cohort_scna_table}
\alias{example_cohort_scna_table}
\title{Example Cohort-level Somatic Copy Number Alteration Table}
\format{
A data frame with 126147 rows and 5 variables:
\describe{
\item{chr}{chromosome the segment is located in}
\item{start}{start position of the segment}
\item{end}{end position of the segment}
\item{log2ratio}{\ifelse{html}{\out{log<sub>2</sub>}}{\eqn{log_2}} ratio of
the segment}
\item{tumor_id}{ID for the tumor containing the SCNA segment}
}
}
\source{
\url{https://dcc.icgc.org/releases/release_28}
}
\usage{
example_cohort_scna_table
}
\description{
A data set containing the somatic copy number alteration data for 10 randomly
selected samples from TCGA's LAML (Acute Myeloid Leukemia) cohort
}
\keyword{datasets}
