% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{fips_codes_for_states_and_dc}
\alias{fips_codes_for_states_and_dc}
\title{Federal Information Processing Standards (FIPS) codes for all states and DC.}
\format{
A data frame with three variables: \code{state_name},
\code{state_fips_code}, and \code{state_abbreviation}.
}
\usage{
fips_codes_for_states_and_dc
}
\description{
Federal Information Processing Standards (FIPS) codes for all states and DC.
}
\keyword{datasets}
