% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/gdoc_to_md.R
\name{gdoc_to_md}
\alias{gdoc_to_md}
\title{Convert Google Doc HTML output to markdown}
\usage{
gdoc_to_md(input_html, output_file = NULL, ...)
}
\arguments{
\item{input_html}{an HTML file retrieved via
download_file(..., download_type='text/html')}

\item{output_file}{a filename for markdown output (optional)}

\item{...}{options (as a character string) to be passed to pandoc via
\code{\link[rmarkdown]{pandoc_convert}}}
}
\value{
If no \code{output_file} is specified, the markdown output is
returned as a character. If \code{output_file} is specified, the markdown is
written to this file and also returned invisibly.
}
\description{
Converts the HTML exported by from a Google document (retrieved
by \code{download_file(..., download_type='text/html')} to a markdown file,
pre-processing to deal with Google Doc's particular formatting.
}
\seealso{
\code{\link{download_file}}, \code{\link[rmarkdown]{pandoc_convert}}
}

