% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geocodio.R
\name{geocode_components}
\alias{geocode_components}
\title{Geocode a single address}
\usage{
geocode_components(
  street,
  city,
  state,
  postal_code,
  fields = "census2010",
  ...
)
}
\arguments{
\item{street, city, state, postal_code}{address components}

\item{fields}{vector of additional fields to return with query results. Note
that these count as extra lookups and impact your dailu quota/costs. See
\href{https://geocod.io/docs/#fields}{the official documentation} for more
information on costs/pricing.}

\item{...}{additional arguments passed to
\code{\link[rgeocodio:gio_geocode_components]{rgeocodio::gio_geocode_components()}}}
}
\value{
a \code{\link[tibble:tibble]{tibble::tibble()}} containing geocodio data per the request
}
\description{
Geocode a single address
}
