% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/comments.R
\name{get_comment}
\alias{get_comment}
\title{retrieve a comment's metadata and content}
\usage{
get_comment(token, file_id, comment_id, simplify = FALSE, ...)
}
\arguments{
\item{token}{a token, generated with \code{\link{driver_connect}}.}

\item{file_id}{the ID of a file - see \code{\link{file_metadata}} for further discussion.}

\item{comment_id}{the ID of a comment, which can be easily retrieved with \code{\link{list_comments}}}

\item{simplify}{whether or not to perform some (small) simplification of the returned
list, to make it less nested, headachey and impossible to read. Set to FALSE by default.}

\item{...}{further arguments to pass to httr's GET}
}
\description{
grabs the metadata and content (both as HTML and plaintext) of a specified
comment to a Google Drive file.
}
\seealso{
\code{\link{list_comments}} to retrieve comment metadata (including the comment
IDs) for an entire file.
}

