% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adrc.R
\name{get_drug_codes}
\alias{get_drug_codes}
\title{Lookup Drug Codes by Name from Kegg.jp}
\usage{
get_drug_codes(drug_name)
}
\arguments{
\item{drug_name}{The name of the drug to lookup.}
}
\value{
A tibble listing drug codes and corresponding names.
}
\description{
Lookup Drug Codes by Name from Kegg.jp
}
\examples{
get_drug_codes("acetaminophen")
}
\references{
\url{https://www.kegg.jp/kegg/drug/}
}
