% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.r
\name{get_file}
\alias{get_file}
\title{Get a file from drive}
\usage{
get_file(file_id, token = get_access_cred())
}
\arguments{
\item{file_id}{the ID for the file}

\item{token}{a valid OAuth2.0 token}
}
\value{
file resource
}
\description{
Get a file from drive
}

