% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.r
\name{get_file_from_title}
\alias{get_file_from_title}
\title{Get a file using file title}
\usage{
get_file_from_title(title, token = get_access_cred())
}
\arguments{
\item{title}{file title}

\item{token}{a valid OAuth2.0 token}
}
\value{
file resource
}
\description{
Get a file using file title
}

