% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/revisions.R
\name{get_revision}
\alias{get_revision}
\title{Get the metadata of a single revision of a Google Drive file}
\usage{
get_revision(token, file_id, rev_id, simplify = FALSE, ...)
}
\arguments{
\item{token}{a token, generated with \code{\link{driver_connect}}.}

\item{file_id}{the ID of a file - see \code{\link{file_metadata}} for further discussion.}

\item{rev_id}{the ID of a revision of that file.}

\item{simplify}{whether or not to perform some (small) simplification of the returned
list, to make it less nested, headachey and impossible to read. Set to FALSE by default.}

\item{...}{further arguments to pass to httr's GET.}
}
\description{
retrieves the metadata associated with a particular revision of a specific
Google Drive file.
}

