% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/breadcrumbs.R
\name{insert_into_breadcrumb_database_table}
\alias{insert_into_breadcrumb_database_table}
\title{Insert into Breadcrumb Database Table}
\usage{
insert_into_breadcrumb_database_table(
  con,
  breadcrumb_database_table,
  monthly_breadcrumb_files
)
}
\arguments{
\item{con}{a DBIConnection object, as returned by \code{dbConnect()}.}

\item{breadcrumb_database_table}{name of the database table for insertion}

\item{monthly_breadcrumb_names}{names of the monthly breadcrumb files}
}
\value{
number of records affected
}
\description{
Insert into Breadcrumb Database Table
}
\examples{
\dontrun{
ris_storage_path("Reports") \%>\%
  collate_daily_breadcrumb_metadata() \%>\%
  write_monthly_breadcrumbs_for_year(2015) -> monthly_breadcrumb_files

con <- DBI::dbConnect(
  RMariaDB::MariaDB(),
  default.file = normalizePath("~./.my.cnf"),
  group = "spark.myftp.org"
)

create_empty_breadcrumb_database_table(
  con, 2015
) -> breadcrumb_database_table

insert_into_breadcrumb_database_table(
  con, breadcrumb_database_table, monthly_breadcrumb_files
)

DBI::dbDisconnect(con)
}
}
