% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crumbs.R
\name{insert_into_crumb_table}
\alias{insert_into_crumb_table}
\title{Insert into Breadcrumb Database Table}
\usage{
insert_into_crumb_table(con, crumb_table, monthly_crumb_files)
}
\arguments{
\item{con}{a DBIConnection object, as returned by \code{dbConnect()}.}

\item{crumb_table}{name of the database table for insertion}

\item{monthly_crumb_files}{names of the monthly breadcrumb files}
}
\value{
number of records affected
}
\description{
Insert into Breadcrumb Database Table
}
\examples{
\dontrun{
ris_storage_path("Reports") \%>\%
  collate_daily_crumb_metadata() \%>\%
  write_monthly_crumbs_for_year(2015) -> monthly_crumb_files

con <- DBI::dbConnect(
  RMariaDB::MariaDB(),
  default.file = normalizePath("~./.my.cnf"),
  group = "spark.myftp.org"
)

create_crumb_table(
  con, 2015
) -> crumb_table

insert_into_crumb_table(
  con, crumb_table, monthly_crumb_files
)

DBI::dbDisconnect(con)
}
}
