% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unsorted.R
\name{invLogit}
\alias{invLogit}
\title{Inverse Logit Transformation of Vector}
\usage{
invLogit(x)
}
\arguments{
\item{x}{vector}
}
\value{
vector
}
\description{
Inverse Logit Transformation of Vector
}
\examples{
invLogit(c(.2, .4, 1, 0))
}
