% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{KEGG_cancer_pathways}
\alias{KEGG_cancer_pathways}
\title{KEGG "Pathways in cancer"-related Pathways - Gene Sets}
\format{
list containing 21 vectors of gene symbols. Each vector corresponds
to a pathway.
}
\usage{
KEGG_cancer_pathways
}
\description{
A list containing the genes involved in each Homo sapiens KEGG "Pathways in
cancer" (hsa05200)-related Pathways. Each element is a vector of gene symbols
located in the given pathway. Names correspond to the KEGG ID of the pathway.
\emph{Generated on Nov 17, 2020.}
}
\seealso{
\code{\link{KEGG_cancer_pathways_descriptions}} for descriptions of
KEGG "Pathways in cancer"-related pathways.
}
\keyword{datasets}
