% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{KEGG_cancer_pathways_descriptions}
\alias{KEGG_cancer_pathways_descriptions}
\title{KEGG "Pathways in cancer"-related Pathways - Descriptions}
\format{
A data frame with 21 rows and 2 variables:
\describe{
\item{id}{KEGG pathway ID}
\item{description}{KEGG pathway description}
}
}
\usage{
KEGG_cancer_pathways_descriptions
}
\description{
A data frame containing descriptions for KEGG "Pathways in cancer"
(hsa05200)-related pathways.
\emph{Generated on Sep 6, 2020.}
}
\keyword{datasets}
