% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adrc.R
\name{kegg_drug_code}
\alias{kegg_drug_code}
\title{Lookup Drug Code(s) from Name via Kegg.jp}
\usage{
kegg_drug_code(name, formal = "")
}
\arguments{
\item{name}{drug name}

\item{formal}{formal Kegg name}
}
\value{
character vector of drug code(s)
}
\description{
Lookup Drug Code(s) from Name via Kegg.jp
}
\note{
Searching via name is often one-to-many. Use \code{kegg_drug_info()} to get
names associated with multiple codes. Results will be filtered to match
formal name if supplied.
}
\examples{
## lookup all associated codes
kegg_drug_code("acetaminophen")

## lookup code for formal name
kegg_drug_code(
  "acetaminophen",
  "Acetaminophen (JP18/USP); Paracetamol (INN); Tylenol (TN)"
)

}
\references{
\url{https://www.kegg.jp/kegg/drug/}
}
