% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adrc.R
\name{kegg_drug_info}
\alias{kegg_drug_info}
\title{Lookup Drug Information from Kegg.jp}
\usage{
kegg_drug_info(x)
}
\arguments{
\item{x}{either the drug name or code}
}
\value{
A tibble listing drug codes and corresponding names.
}
\description{
Lookup Drug Information from Kegg.jp
}
\examples{
kegg_drug_info("acetaminophen")

}
\references{
\url{https://www.kegg.jp/kegg/drug/}
}
