% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adrc.R
\name{kegg_drug_name}
\alias{kegg_drug_name}
\title{Lookup Drug Name from Code via Kegg.jp}
\usage{
kegg_drug_name(code)
}
\arguments{
\item{code}{drug code}
}
\value{
character descriptor of drug code
}
\description{
Lookup Drug Name from Code via Kegg.jp
}
\examples{
kegg_drug_name("d00217")
}
\references{
\url{https://www.kegg.jp/kegg/drug/}
}
