% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.r
\name{list_all_files}
\alias{list_all_files}
\title{List all files}
\usage{
list_all_files(order_by = NULL, token = get_access_cred())
}
\arguments{
\item{order_by}{a comma-separated list of sort keys}

\item{token}{a valid OAuth2.0 token}
}
\value{
file resources
}
\description{
List all files
}

