% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/comments.R
\name{list_comments}
\alias{list_comments}
\title{retrieve the metadata and content of all comments to a file}
\usage{
list_comments(file_id, simplify = FALSE, ...)
}
\arguments{
\item{file_id}{the ID of a file - or the full URL for accessing it via your browser.
See \code{\link{file_metadata}} for further discussion.}

\item{simplify}{whether or not to perform some (small) simplification of the returned
list, to make it less nested, headachey and impossible to read. Set to FALSE by default.}

\item{...}{further arguments to pass to httr's GET}
}
\description{
grabs the metadata and content (both as HTML and plaintext) of each comment
in a specified Google Drive file.
}
\seealso{
\code{\link{list_comments}} to retrieve comment metadata (including the comment
IDs) for an entire file.
}

