% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/files.R
\name{list_files}
\alias{list_files}
\title{Retrieve the metadata for all files}
\usage{
list_files(token, max_results = 100, page_token = NULL, simplify = FALSE,
  ...)
}
\arguments{
\item{token}{a token, generated with \code{\link{driver_connect}}.}

\item{max_results}{the maximum number of results to return; any number between 1 and 1000.
Set to 100 by default.}

\item{page_token}{in the event that the requested files are split over multiple pages,
each object returned from \code{list_files} will contain an element named "nextPageToken".
Plugging this into the \code{page_token} parameter provides for query continuation.}

\item{simplify}{whether or not to perform some (small) simplification of the returned
list, to make it less nested, headachey and impossible to read. Set to FALSE by default.}

\item{...}{further arguments to pass to httr's GET.}
}
\description{
\code{list_files} allows an authenticated user to retrieve the metadata
associated with each file they have access to. For the metadata for a single file, see
\code{\link{file_metadata}}.
}

