% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.r
\name{list_files_from_title}
\alias{list_files_from_title}
\title{List files using file title}
\usage{
list_files_from_title(title, token = get_access_cred())
}
\arguments{
\item{title}{file title}

\item{token}{a valid OAuth2.0 token}
}
\value{
req
}
\description{
List files using file title
}

