% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/replies.R
\name{list_replies}
\alias{list_replies}
\title{Get all replies to a comment}
\usage{
list_replies(token, file_id, comment_id, simplify = FALSE, ...)
}
\arguments{
\item{token}{a token, generated with \code{\link{driver_connect}}.}

\item{file_id}{the ID of a file - see \code{\link{file_metadata}} for further discussion.}

\item{comment_id}{the ID of a comment, which can be found in the output of \code{\link{list_comments}}}

\item{simplify}{whether or not to perform some (small) simplification of the returned
list, to make it less nested, headachey and impossible to read. Set to FALSE by default.}

\item{...}{further arguments to pass to httr's GET}
}
\description{
retrieves the metadata and text of all replies to a comment in a
Google Drive file.
}
\seealso{
\code{\link{delete_reply}} for deleting a reply, and \code{\link{get_reply}}
for retrieving a specific reply.
}

