% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/revisions.R
\name{list_revisions}
\alias{list_revisions}
\title{list revisions of a Google Drive file}
\usage{
list_revisions(token, file_id, simplify = FALSE, ...)
}
\arguments{
\item{token}{a token, generated with \code{\link{driver_connect}}.}

\item{file_id}{the ID of a file - or the full URL for accessing it via your browser.
See \code{\link{file_metadata}} for further discussion.}

\item{simplify}{whether or not to perform some (small) simplification of the returned
list, to make it less nested, headachey and impossible to read. Set to FALSE by default.}

\item{...}{further arguments to pass to httr's GET.}
}
\description{
retrieves the metadata associated with each revision of a specific
Google Drive file.
}

