% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rds.r
\name{load_rds}
\alias{load_rds}
\title{Load an object from Rds file from drive}
\usage{
load_rds(title, token = get_access_cred())
}
\arguments{
\item{title}{file title}

\item{token}{a valid OAuth2.0 token}
}
\description{
Load an object from Rds file from drive
}

