% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unsorted.R
\name{Logit}
\alias{Logit}
\title{Logit Transformation of Vector}
\usage{
Logit(x)
}
\arguments{
\item{x}{vector}
}
\value{
vector
}
\description{
Logit Transformation of Vector
}
\examples{
Logit(c(.2, .4, 1, 0))

}
