% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_utils.R
\name{lower_triangle_factors}
\alias{lower_triangle_factors}
\title{Vector to lower diagonal of matrix formed from factor by factor}
\usage{
lower_triangle_factors(a, VarNames = c("Var1", "Var2"))
}
\arguments{
\item{a}{vector}

\item{VarNames}{character vector of length 2}
}
\value{
data.frame with two columns named according to \code{VarNames}
}
\description{
Formed by unique(vetor) %times% unique(vector)
}
\details{
Useful for creating heatmaps using only lower diagonal.
}
\examples{
lower_triangle_factors(rep(1:3, 10))
}
