% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crumbs.R
\docType{data}
\name{master_crumb_template}
\alias{master_crumb_template}
\title{Master Breadcrumb Template}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 0 rows and 23 columns.
}
\usage{
master_crumb_template
}
\description{
Daily breadcrumb formatting is inconsistent where the fields provided
differ across files. Notably some contain a datetime field where the norm is
to split date and time into separate fields. This tibble represents a base
template to store all known breadcrumb data in a common format.
}
\keyword{datasets}
