% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{MCR_table}
\alias{MCR_table}
\title{Table of Pan-Cancer Minimal Common Regions}
\format{
A data frame with 165 rows and 5 variables:
\describe{
\item{chr}{chromosome the MCR is located in}
\item{start}{start position of the MCR}
\item{end}{end position of the MCR}
\item{MCR_type}{the type ("Amp" or "Del") of the MCR peak}
\item{SCNA_density}{SCNA per Mb within the MCR}
}
}
\source{
\url{https://pubmed.ncbi.nlm.nih.gov/23132910/}
}
\usage{
MCR_table
}
\description{
A data set containing the minimal common regions (MCRs) across all cancer
types studied in Kim TM, Xi R, Luquette LJ, Park RW, Johnson MD, Park PJ.
Functional genomic analysis of chromosomal aberrations in a compendium of
8000 cancer genomes. Genome Res. 2013;23(2):217-27.
}
\keyword{datasets}
