% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{metapredictor_model}
\alias{metapredictor_model}
\title{Random Forest Model for Coding Impact Metaprediction}
\format{
model object
}
\usage{
metapredictor_model
}
\description{
A Random Forest model object for metaprediction of coding variants' impact,
using 12 impact scores from different coding impact predictors. The model was
trained on 711 coding variants, with 10-folds repeated 3 times cross-validation.
}
\keyword{datasets}
