% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/array_coda.R
\name{miniclo_array}
\alias{miniclo_array}
\title{Closure Operation applied to array on margin}
\usage{
miniclo_array(a, samples = 1, parts = 2)
}
\arguments{
\item{a}{multidimensional array}

\item{samples}{index of dimension of \code{x} that represents samples}

\item{parts}{index of dimension of \code{x} that represents parts (e.g., compositional variables)}
}
\value{
array
}
\description{
Array version of \code{\link{miniclo}}
}
\examples{
a <- array(1:100, dim=c(10, 5, 2))
miniclo_array(a)
}
