% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{MTL_submodel_descriptions}
\alias{MTL_submodel_descriptions}
\title{MTL Sub-model Descriptions}
\format{
A data frame with 21 rows and 2 variables:
\describe{
\item{short_name}{short name for the cancer type}
\item{description}{description of the cancer type}
}
}
\usage{
MTL_submodel_descriptions
}
\description{
A data frame containing descriptions for all sub-models of the MTL model.
}
\seealso{
\code{\link{TCGA_MTL_fit}} for the MTL model.
}
\keyword{datasets}
