% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/big_book.R
\name{publish_big_book_of_drives_notes}
\alias{publish_big_book_of_drives_notes}
\title{Publish Big Book of Drives Notes}
\usage{
publish_big_book_of_drives_notes(
  server = "wuit-s-12552.accounts.ad.wustl.edu",
  user = Sys.getenv("WUSTL_KEY_USER"),
  pass = Sys.getenv("WUSTL_KEY_PASS")
)
}
\arguments{
\item{server}{FQDN of server}

\item{user}{WUSTL Key username}

\item{pass}{WUSTL Key password}
}
\description{
Publish Big Book of Drives Notes
}
\note{
After building the Big Book of Drives Notes, run this function while
connected to msvpn.wusm.wustl.edu to publish on the server.
}
