% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crumbs.R
\name{read_collated_crumb}
\alias{read_collated_crumb}
\title{Read a Collated Breadcrumb}
\usage{
read_collated_crumb(x)
}
\arguments{
\item{x}{collated breadcrumb file}
}
\value{
tibble representation of collated breadcrumb contents
}
\description{
Read a Collated Breadcrumb
}
\note{
Read is based on \code{master_crumb_template} format.
}
\examples{
\dontrun{
read_collated_crumb("breadcrumb_2015-01-01_to_2015-02-01.csv")
}

}
