% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/breadcrumbs.R
\name{read_daily_breadcrumb}
\alias{read_daily_breadcrumb}
\title{Read Daily Breadcrumb}
\usage{
read_daily_breadcrumb(file, verbose = FALSE)
}
\arguments{
\item{file}{breadcrumb to read}

\item{verbose}{whether to print header type messages}
}
\value{
tibble of properly formatted breadcrumb data according to
\code{master_breadcrumb_template}
}
\description{
Read Daily Breadcrumb
}
\note{
Breadcrumb files are inconsistently formatted. Known issues include
that some are comma delimited and some semicolon, some contain datetimes and
some contain dates and times, some use '' and some use '.' as missing, and
some included fields are inconsistent from one file to the next - especially
the datetimes vs date and time issue. This function explicitly reads the
file according to the file header in an effort to detect data issues.
}
\examples{
\dontrun{
read_daily_breadcrumb("Breadcrumb_07Nov2017.csv")
}

}
