% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/breadcrumbs.R
\name{read_daily_breadcrumb_header}
\alias{read_daily_breadcrumb_header}
\title{Read Daily Breadcrumb Header}
\usage{
read_daily_breadcrumb_header(x)
}
\arguments{
\item{x}{breadcrumb file from which to extract header}
}
\value{
character representation of the breadcrumb header
}
\description{
Read Daily Breadcrumb Header
}
\note{
Used to determine breadcrumb read in \code{read_daily_breadcrumb()}.
}
\examples{
\dontrun{
read_daily_breadcrumb_header("Breadcrumb_07Nov2017.csv")
}

}
