% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redcap.R
\name{redcap_logical}
\alias{redcap_logical}
\title{REDCap Logical}
\usage{
redcap_logical(x)
}
\arguments{
\item{x}{R logical to cast as REDCap API logical}
}
\value{
an R logical represented as a named list of lowercase character
}
\description{
REDCap Logical
}
\examples{
\dontrun{
a <- TRUE
redcap_logical(a)

b <- FALSE
redcap_logical(b)

c <- "not gonna work"
redcap_logical(c)
}
}
