% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redcap.R
\name{redcap_read}
\alias{redcap_read}
\title{Retrieve credentials from a csv file and read records from a REDCap project}
\usage{
redcap_read(project_id, path_credential = "~/.REDCapR", ...)
}
\arguments{
\item{project_id}{The ID assigned to the project withing REDCap. This allows
the user to store tokens to multiple REDCap projects in one file.}

\item{path_credential}{The file path to the CSV containing the credentials.}

\item{...}{Additional arguments passed to
\code{\link[REDCapR:retrieve_credential]{REDCapR::retrieve_credential_local()}} and \code{\link[REDCapR:redcap_read]{REDCapR::redcap_read()}}.}
}
\value{
An R \code{\link[base:data.frame]{base::data.frame()}} of the desired records and columns.
}
\description{
Retrieve credentials from a csv file and read records from a REDCap project
}
\examples{
\dontrun{
redcap_read(6648)
}
}
