% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/breadcrumbs.R
\name{remove_monthly_breadcrumbs}
\alias{remove_monthly_breadcrumbs}
\title{Remove Monthly Breadcrumb Files}
\usage{
remove_monthly_breadcrumbs(monthly_breadcrumb_files)
}
\arguments{
\item{monthly_breadcrumb_files}{names of the monthly breadcrumb files to
collate (i.e., return value from \code{write_monthly_breadcrumbs_for_year()})}
}
\description{
Remove Monthly Breadcrumb Files
}
\examples{
\dontrun{
# silly example to write breadcrumbs and the immediately delete them
ris_storage_path("Reports") \%>\%
  collate_daily_breadcrumb_metadata() \%>\%
  write_monthly_breadcrumbs_for_year(2015) -> monthly_breadcrumb_files

monthly_breadcrumb_files \%>\%
  collate_monthly_breadcrumbs()

monthly_breadcrumb_files \%>\%
  remove_monthly_breadcrumbs()
}
}
