% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crumbs.R
\name{remove_monthly_crumbs}
\alias{remove_monthly_crumbs}
\title{Remove Monthly Breadcrumb Files}
\usage{
remove_monthly_crumbs(monthly_crumb_files)
}
\arguments{
\item{monthly_crumb_files}{names of the monthly breadcrumb files to
collate (i.e., return value from \code{write_monthly_crumbs_for_year()})}
}
\description{
Remove Monthly Breadcrumb Files
}
\examples{
\dontrun{
# silly example to write breadcrumbs and the immediately delete them
ris_storage_path("Reports") \%>\%
  collate_daily_crumb_metadata() \%>\%
  write_monthly_crumbs_for_year(2015) -> monthly_crumb_files

monthly_crumb_files \%>\%
  collate_monthly_crumbs()

monthly_crumb_files \%>\%
  remove_monthly_crumbs()
}
}
