% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/breadcrumbs.R
\name{ris_storage_path}
\alias{ris_storage_path}
\title{Path to Research Infrastructure Services Storage}
\usage{
ris_storage_path(path = "")
}
\arguments{
\item{path}{additional path beyond root mapping}
}
\value{
the desired path
}
\description{
Path to Research Infrastructure Services Storage
}
\examples{
\dontrun{
breadcrumb_directory <- ris_storage_path("Reports")
}
}
