% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/box.R
\name{roe_box_ftp_upload}
\alias{roe_box_ftp_upload}
\title{Upload content to the Roe Lab folder on Box via FTP}
\usage{
roe_box_ftp_upload(file, path, user, pass, ...)
}
\arguments{
\item{file}{name or path of the local file to be uploaded.}

\item{path}{the path to which the content is to be uploaded.}

\item{user}{Box username (i.e., WashU email)}

\item{pass}{unique password for external applications. Created at
\url{https://wustl.app.box.com/account}}

\item{...}{additional arguments passed to \code{\link[RCurl:ftpUpload]{RCurl::ftpUpload()}}}
}
\value{
The result of the \code{\link[RCurl]{curlPerform}} call.
}
\description{
Upload content to the Roe Lab folder on Box via FTP
}
\examples{
\dontrun{
file <- tempfile("roe_box_ftp_upload_test", fileext = ".txt")
writeLines(c(
  "roelabr::roe_box_ftp_upload()",
  sprintf("uploaded this at \%s.", Sys.time())
), file)
path <- file.path("test_directory", basename(file))
user <- Sys.getenv("WUSTL_BOX_USER")
pass <- Sys.getenv("WUSTL_BOX_PASS")
roe_box_ftp_upload(file, path, user, pass)
}

}
\references{
\url{https://wustl.app.box.com/services/box_ftp_server/}
}
