% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geocodio.R
\name{roe_geocode_components}
\alias{roe_geocode_components}
\title{Geocode a single address}
\usage{
roe_geocode_components(
  street,
  city,
  state,
  postal_code,
  fields = "census2010",
  ...
)
}
\arguments{
\item{street, city, state, postal_code}{address components}

\item{fields}{vector of additional fields to return with query results. Note
that these count as extra lookups and impact your dailu quota/costs. See
[the official documentation](https://geocod.io/docs/#fields) for more
information on costs/pricing.}

\item{...}{additional arguments passed to
[rgeocodio::gio_geocode_components()]}
}
\value{
a [tibble::tibble()] containing geocodio data per the request
}
\description{
Geocode a single address
}
