% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/census.R
\name{roe_get_census_urban_rural_per_block_group_2010}
\alias{roe_get_census_urban_rural_per_block_group_2010}
\title{Retrieve Urban and Rural (p2) Summary File 1 (sf1) Census data from the 2010 decennial census}
\usage{
roe_get_census_urban_rural_per_block_group_2010(state_fips_code)
}
\arguments{
\item{state_fips_code}{Two-digit state code (e.g., "01" == "Alabama").}
}
\value{
a tibble with FIPS codes and P2 variables by tract group for the desired state
}
\description{
Retrieve Urban and Rural (p2) Summary File 1 (sf1) Census data from the 2010 decennial census
}
\details{
Decennial Census (2010) API Documentation
\url{https://www.census.gov/data/developers/data-sets/decennial-census.html}

URBAN AND RURAL variables listed in:
\url{https://api.census.gov/data/2010/dec/sf1/variables.html}
\tabular{llllllll}{
\strong{Name} \tab \strong{Label} \tab \strong{Concept} \tab \strong{Required} \tab \strong{Attributes} \tab \strong{Limit} \tab \strong{Predicate Type} \tab \strong{Group} \cr
P002001 \tab Total	                                \tab URBAN AND RURAL \tab not required \tab P002001ERR \tab 0 \tab int  \tab P2 \cr
P002002 \tab Total!!Urban	                        \tab URBAN AND RURAL \tab not required \tab            \tab 0 \tab int  \tab P2 \cr
P002003 \tab Total!!Urban!!Inside urbanized areas  \tab URBAN AND RURAL \tab not required \tab            \tab 0 \tab int  \tab P2 \cr
P002004 \tab Total!!Urban!!Inside urban clusters	  \tab URBAN AND RURAL \tab not required \tab            \tab 0 \tab int  \tab P2 \cr
P002005 \tab Total!!Rural	                        \tab URBAN AND RURAL \tab not required \tab            \tab 0 \tab int  \tab P2 \cr
P002006 \tab Total!!Not defined for this file	    \tab URBAN AND RURAL \tab not required \tab            \tab 0 \tab int  \tab P2
}
}
\examples{
\dontrun{
roe_get_census_urban_rural_per_block_group_2010("01")
}

}
