% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redcap.R
\name{roe_get_redcap_sas_export}
\alias{roe_get_redcap_sas_export}
\title{Export Full REDCap Databases to SAS}
\usage{
roe_get_redcap_sas_export(
  token,
  filename = "roe_redcap_sas_export",
  redcap_uri = "https://redcap.wustl.edu/redcap/api/"
)
}
\arguments{
\item{token}{The user-specific string that serves as the password for a project.}

\item{filename}{The name to be used for all output files (.csv, .sas, .log, and .sas7bdat). Must be a valid SAS name.}

\item{redcap_uri}{The URI (uniform resource identifier) of the REDCap project.}
}
\description{
The SAS export capability of REDCap v9.5 is broken such that the
 generated SAS script has voluminous syntactical errors. This function
attempts to replicate the intended functionality of REDCap by downloading the
 data to a csv file and generating a syntactically correct SAS import script.
  If sas.exe is found on the path, the script is executed in SAS batch mode
  to produce the desired sas7bdat file.
}
\examples{
\dontrun{
token <- REDCapR::retrieve_credential_local("~/.REDCapR", 7842)$token
roe_get_redcap_sas_export(token, "static")

token <- REDCapR::retrieve_credential_local("~/.REDCapR", 6785)$token
roe_get_redcap_sas_export(token, "mother")
}
}
