% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{roe_timestamp_filename}
\alias{roe_timestamp_filename}
\title{Append a Timestamp to the End of a File Name}
\usage{
roe_timestamp_filename(x)
}
\arguments{
\item{x}{file name to which a time stamp string should be appended.}
}
\value{
file name with time stamp string appended
}
\description{
Append a Timestamp to the End of a File Name
}
\examples{
roe_timestamp_filename("roe_redcap_sas_export")
}
