% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roelabr-package.R
\docType{package}
\name{roelabr-package}
\alias{roelabr}
\alias{roelabr-package}
\title{roelabr: A collection of functions for the Roe Lab at Washington University}
\description{
A collection of functions for the Roe Lab at Washington University. Having a central location for commonly utilized code is good practice.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/RoeLabWustl/roelabr}
  \item Report bugs at \url{https://github.com/RoeLabWustl/roelabr/issues}
}

}
\author{
\strong{Maintainer}: Matthew Schuelke \email{matt@themadstatter.com} (\href{https://orcid.org/0000-0001-5755-1725}{ORCID})

Authors:
\itemize{
  \item Julie Wisch \email{julie.wisch@wustl.edu} (\href{https://orcid.org/0000-0003-3624-2784}{ORCID})
}

}
\keyword{internal}
