% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampling.R
\name{sample_uniform}
\alias{sample_uniform}
\alias{rUnifSphere}
\alias{rUnifSimplex}
\title{Sample Uniformily from various objects centered at the origin}
\usage{
rUnifSphere(D, radius, n_samples = 1, shell_only = FALSE)

rUnifSimplex(D, n_samples = 1)
}
\arguments{
\item{D}{dimension of object}

\item{radius}{radius of sphere to sample from}

\item{n_samples}{number of samples to produce}

\item{shell_only}{if TRUE samples from shell of object only, if false
samples from entire region of object}
}
\value{
matrix of dimension D x n_samples
}
\description{
Sample Uniformily from various objects centered at the origin
}
