% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rds.r
\name{save_rds}
\alias{save_rds}
\title{Save an object as Rds file to drive}
\usage{
save_rds(obj, title = NULL, email = NULL, token = get_access_cred())
}
\arguments{
\item{obj}{object to be uploaded}

\item{title}{file title}

\item{email}{email address to add to permission list}

\item{token}{a valid OAuth2.0 token}
}
\value{
file id
}
\description{
Save an object as Rds file to drive
}

