% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{specific_thresholds}
\alias{specific_thresholds}
\title{Tumor type specific probability thresholds}
\format{
vector with 21 elements
}
\usage{
specific_thresholds
}
\description{
Driver gene probability thresholds for all 21 cancer types (submodels).
}
\seealso{
\code{\link{TCGA_MTL_fit}} for the Multi-Task Learning model.
}
\keyword{datasets}
