% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/array_munging.R
\name{split.along.dim}
\alias{split.along.dim}
\title{Split Multidimeninal array to a list along a specified dimension}
\usage{
\method{split}{along.dim}(a, n)
}
\arguments{
\item{a}{array}

\item{n}{dimention of array to split}
}
\value{
List of array of 1 less dimension than a
}
\description{
Split Multidimeninal array to a list along a specified dimension
}
\examples{
a <- array(0, dim=4:6)
split.along.dim(a, 3)
}
\references{
https://stackoverflow.com/questions/20198751/three-dimensional-array-to-list
}
