% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/array_munging.R
\name{spread_array}
\alias{spread_array}
\title{Spread data.frame}
\usage{
spread_array(data, value, ...)
}
\arguments{
\item{data}{a dataframe with integer indexed positions in an array}

\item{value}{data to be spread into the array}

\item{...}{unquoted dimnames used to build array}
}
\value{
multidimensional array
}
\description{
Spread data.frame
}
\details{
make sure all colnames supplied to ... are consecutive interger indexed positions
}
\examples{
a <- array(1:100, dim=c(10, 5, 2))
ga <- gather_array(a, sequence, A, B, C)
head(ga)
spread_array(ga, sequence, A, B, C)
}
\seealso{
gather_array
}
