% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_shortcuts.R
\name{summarise_posterior}
\alias{summarise_posterior}
\title{Shortcut for summarize variable with quantiles and mean}
\usage{
summarise_posterior(data, var, ...)
}
\arguments{
\item{data}{tidy data frame}

\item{var}{variable name (unquoted) to be summarised}

\item{...}{other expressions to pass to summarise}
}
\value{
data.frame
}
\description{
Shortcut for summarize variable with quantiles and mean
}
\examples{
# library(dplyr)
# d <- data.frame("a" = c(1:10), "b"=rep(c(1,2), 5)) \%>\%
#   summarise_posterior(a, mean.b = mean(b))
}
