% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{TCGA_MTL_fit}
\alias{TCGA_MTL_fit}
\title{Multi-Task Learning Model for Predicting Cancer Driverness of Genes}
\format{
MTL model object
}
\usage{
TCGA_MTL_fit
}
\description{
A Multi-Task Learning (MTL) classification model object for determining
cancer driverness of genes based on 26 features. The model was trained using
TCGA data (obtained from ICGC release 28) with lasso regularization. It contains
21 sub-models for different cancer types.
}
\keyword{datasets}
