% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/files.R
\name{trash_file}
\alias{trash_file}
\title{move a file to the trash}
\usage{
trash_file(file_id, ...)
}
\arguments{
\item{file_id}{the ID of the file; see \code{\link{file_metadata}} for further
commentary.}

\item{...}{further arguments to pass to httr's POST.}
}
\value{
TRUE if the file was successfully trashed, an error otherwise.
}
\description{
moves a file to Google Drive's "trash" folder, which is automatically
emptied after a set number of days.
}

