% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delete.r
\name{trash_files_from_title}
\alias{trash_files_from_title}
\title{Move file(s) to trash}
\usage{
trash_files_from_title(title, all = FALSE, token = get_access_cred())
}
\arguments{
\item{title}{file title}

\item{all}{whether to move all file(s) having the same title to trash}

\item{token}{a valid OAuth2.0 token}
}
\value{
ID(s) of file(s) moved to trash
}
\description{
Move file(s) to trash
}

