% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/files.R
\name{update_file_metadata}
\alias{update_file_metadata}
\title{update the metadata associated with a Google Drive file}
\usage{
update_file_metadata(token, metadata, ...)
}
\arguments{
\item{token}{a token, generated with \code{\link{driver_connect}}.}

\item{metadata}{a blob of metadata retrieved from \code{\link{get_metadata}}, with your modifications
made to it.}

\item{...}{further arguments to pass to httr's PATCH.}
}
\value{
the new metadata associated with the file, allowing you to confirm the changes took effect.
}
\description{
\code{update_file_metadata} allows you to update the metadata associated
with a file on Google Drive - for example, changing the title or description, or modifying
the permissions.
}

