% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/files.R
\name{update_file_time}
\alias{update_file_time}
\title{update the time a file was viewed}
\usage{
update_file_time(token, file_id, ...)
}
\arguments{
\item{token}{a token, generated with \code{\link{driver_connect}}.}

\item{file_id}{the ID of the file; see \code{\link{file_metadata}} for further
commentary.}

\item{...}{further arguments to pass to httr's POST.}
}
\value{
TRUE if the file was successfully deleted, an error otherwise.
}
\description{
\code{\link{update_file_time}} updates the metadata associated with a specific
file to state that the file was last viewed/modified at [system time].
}

