% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/files.R
\name{upload_file}
\alias{upload_file}
\title{Upload a local file to Google Drive}
\usage{
upload_file(token, file_path, title = NULL, description = NULL, ...)
}
\arguments{
\item{token}{a token, generated with \code{\link{driver_connect}}.}

\item{file_path}{the full path to the file you want to upload}

\item{title}{what to set in the "title" field of the resulting Google Drive file}

\item{description}{what to set in the "description" field.}
}
\value{
a metadata object referring to the uploaded file, which can be used in subsequent requests. Matches the
output format of \code{\link{get_metadata}}
}
\description{
\code{upload_file} allows you to upload a locally-stored file to Google Drive, setting title and
description along the way if you so choose.
}

