% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unsorted.R
\name{vec_to_mat}
\alias{vec_to_mat}
\title{Check if Vector and if so, Convert to Row Matrix}
\usage{
vec_to_mat(x)
}
\arguments{
\item{x}{vector or matrix}
}
\value{
matrix
}
\description{
Check if Vector and if so, Convert to Row Matrix
}
\examples{
vec_to_mat(c(1,2,3))
vec_to_mat(rbind(c(1,2,3), c(1,2,3)))
}
