% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/breadcrumbs.R
\name{write_monthly_breadcrumbs_for_year}
\alias{write_monthly_breadcrumbs_for_year}
\title{Write Monthly Breadcrumbs for Year}
\usage{
write_monthly_breadcrumbs_for_year(daily_breadcrumb_metadata, year)
}
\arguments{
\item{daily_breadcrumb_metadata}{tibble of breadcrumb names and dates from
\code{collate_daily_breadcrumb_metadata()}}

\item{year}{the desired year}
}
\value{
names of the timespan (monthly) breadcrumb files
}
\description{
Write Monthly Breadcrumbs for Year
}
\examples{
\dontrun{
ris_storage_path("Reports") \%>\%
  collate_daily_breadcrumb_metadata() \%>\%
  write_monthly_breadcrumbs_for_year(2015)
}
}
