% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crumbs.R
\name{write_timespan_crumb}
\alias{write_timespan_crumb}
\title{Write a Timespan Breadcrumb}
\usage{
write_timespan_crumb(daily_crumb_metadata, d0, d1)
}
\arguments{
\item{daily_crumb_metadata}{tibble of breadcrumb names and dates from
\code{collate_daily_crumb_metadata()}}

\item{d0}{start date (e.g., '2022-01-01')}

\item{d1}{end date (e.g., '2022-02-01')}
}
\value{
name of the timespan breadcrumb file invisibly
}
\description{
Write a Timespan Breadcrumb
}
\examples{
\dontrun{
ris_storage_path("Reports") \%>\%
  collate_daily_crumb_metadata() \%>\%
  write_timespan_crumb("2022-01-01", "2022-02-01")
}
}
