% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/array_coda.R
\name{array_lr_transforms}
\alias{array_lr_transforms}
\alias{glr_array}
\alias{glrInv_array}
\alias{alr_array}
\alias{alrInv_array}
\alias{ilr_array}
\alias{ilrInv_array}
\alias{clr_array}
\alias{clrInv_array}
\title{Log-Ratios Transforms for Arrays}
\usage{
glr_array(x, V, samples = 1, parts = 2)

glrInv_array(y, V, samples = 1, coords = 2)

alr_array(x, d = dim(x)[parts], samples = 1, parts = 2)

alrInv_array(y, d = dim(y)[coords] + 1, samples = 1, coords = 2)

ilr_array(x, V = NULL, samples = 1, parts = 2)

ilrInv_array(y, V = NULL, samples = 1, coords = 2)

clr_array(x, samples = 1, parts = 2)

clrInv_array(y, samples = 1, coords = 2)
}
\arguments{
\item{x}{multidimensional array in simplex}

\item{V}{transformation matrix (defines transform)}

\item{samples}{index of dimension of \code{x} that represents samples}

\item{parts}{index of dimension of \code{x} that represents parts (e.g., compositional variables)}

\item{y}{multidimensional array in transformed space}

\item{coords}{index of dimension of \code{x} that represents coords (e.g., transformed variables)}

\item{d}{for ALR, which component (integer position) to take as reference
(default is ncol(x)) for alrInv corresponds to column position in untransformed
matrix.}
}
\value{
array
}
\description{
Extension of \code{\link{base_lr_transforms}} to arrays over arbitrary
margins.
}
\examples{
a <- array(1:100, dim=c(10, 5, 2))
a <- miniclo_array(a)
alr_array(a, 2)
ilr_array(a)
clrInv_array(clr_array(a))
}
