% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/complex_coda.R
\name{convert_coda_covariance}
\alias{convert_coda_covariance}
\alias{ilrvar2clrvar}
\alias{clrvar2ilrvar}
\alias{clrvar2varmat}
\alias{ilrvar2varmat}
\title{Convert CoDA covariance matricies between representations}
\usage{
ilrvar2clrvar(Sigma, V)

clrvar2ilrvar(Sigma, V)

clrvar2varmat(Sigma)

ilrvar2varmat(Sigma, V)
}
\arguments{
\item{Sigma}{covariance matrix in specified transformed space}

\item{V}{ILR contrast matrix (i.e., transformation matrix of ILR)}
}
\value{
matrix
}
\description{
\code{ilrvar}, \code{clrvar}, and \code{varmat} (variationi matrix).
}
\examples{
x <- matrix(runif(30),  10, 3)
x <- miniclo(x)
x.ilr <- ilr(x)
V <- create_default_ilr_base(3)
Sigma <- cov(x.ilr)
Sigma.clr <- ilrvar2clrvar(Sigma, V)
clrvar2ilrvar(Sigma.clr, V)
clrvar2varmat(Sigma.clr)
ilrvar2varmat(Sigma, V)
}
