% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{create_metaprediction_score_df}
\alias{create_metaprediction_score_df}
\title{Create Coding Impact Meta-prediction Score Data Frame}
\usage{
create_metaprediction_score_df(annovar_csv_path)
}
\arguments{
\item{annovar_csv_path}{path to ANNOVAR csv output file}
}
\value{
data frame of meta-prediction scores containing 2 cclumns: \describe{
\item{gene_symbol}{ID of the enriched term}
\item{metaprediction_score}{the metapredictor impact score}
}
}
\description{
Create Coding Impact Meta-prediction Score Data Frame
}
\examples{
path2annovar_csv <- system.file("extdata/LGG-US.hg19_multianno.csv",
package = "driveR")
metapred_df <- driveR:::create_metaprediction_score_df(path2annovar_csv)
}
