% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{determine_hotspot_genes}
\alias{determine_hotspot_genes}
\title{Determine Hotspot Containing Genes}
\usage{
determine_hotspot_genes(annovar_csv_path, hotspot_threshold = 5L)
}
\arguments{
\item{annovar_csv_path}{path to 'ANNOVAR' csv output file}

\item{hotspot_threshold}{to determine hotspot genes, the (integer) threshold
for the minimum number of cases with certain mutation in COSMIC (default = 5)}
}
\value{
vector of gene symbols of genes containing hotspot mutation(s)
}
\description{
Determine Hotspot Containing Genes
}
