% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/utilities.R
\name{driver_connect}
\alias{driver_connect}
\title{connect to Google Drive and authorise driver to access your files.}
\usage{
driver_connect(id, secret)
}
\arguments{
\item{id}{the clientID of the client app you've created (see above).}

\item{secret}{the secret of the client app you've created (see above).}
}
\value{
an OAuth token that can be passed into other function calls.
}
\description{
\code{driver_connect} produces an OAuth 2.0 token which can be passed
into other functions in the driver package, enabling driver to access your Google Drive
files on your behalf.

A necessary prerequisite is creating a client app:

- Go to \url{https://console.developers.google.com}
- Create a client app, with http://localhost:1410/ as the redirect URI
- Connect it to the Drive API.
- note the provided clientID and secret.
}
\examples{
\dontrun{
#Use driver_connect to generate a token, pass that token in to list_files
token <- driver_connect("foo", "bar")
}
}

