% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{get_radius_gyration}
\alias{get_radius_gyration}
\title{Calculate Radius of Gyration}
\usage{
get_radius_gyration(
  DF,
  Latitude_Name = "Trip_End_Latitude",
  Longitude_Name = "Trip_End_Longitude"
)
}
\arguments{
\item{DF}{The dataframe that contains all trips that you want to calculate the radius of gyration for}

\item{Latitude_Name}{Character string of the column name in DF that has the trip end latitude location}

\item{Longitude_Name}{Character string of the column name in DF that has the trip end longitude location}
}
\value{
Rg a single numeric value that gives the radius of gyration for the supplied dataframe in kilometers
}
\description{
Calculate Radius of Gyration
}
\references{
\url{https://blog.jlevente.com/radius-of-gyration/}
\url{https://alzres.biomedcentral.com/articles/10.1186/s13195-021-00852-1}
}
\author{
Julie Wisch
}
