% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geocodio.R
\name{roe_geocode_components}
\alias{roe_geocode_components}
\title{Geocode a single address}
\usage{
roe_geocode_components(street, city, state, fivedigzipcode, ...)
}
\arguments{
\item{street}{address components}

\item{city}{address components}

\item{state}{address components}
}
\description{
\strong{You can also geocode intersections.} Just specify the two streets that you want to
geocode in your query.
}
\examples{
gio_geocode_components("1109 N Highland St", "Arlington", "VA")
gio_geocode_components("1109 N Highland St", "Arlington", "VA",
                       fields=c("census", "stateleg"))
}
