% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delete.r
\name{trash_file}
\alias{trash_file}
\title{Move a file to trash}
\usage{
trash_file(file_id, token = get_access_cred())
}
\arguments{
\item{file_id}{the ID of the file}

\item{token}{a valid OAuth2.0 token}
}
\value{
ID of the trashed file
}
\description{
Move a file to trash
}

