% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/files.R
\name{untrash_file}
\alias{untrash_file}
\title{move a file out of the trash}
\usage{
untrash_file(token, file_id, ...)
}
\arguments{
\item{token}{a token, generated with \code{\link{driver_connect}}.}

\item{file_id}{the ID of the file; see \code{\link{file_metadata}} for further
commentary.}

\item{...}{further arguments to pass to httr's POST.}
}
\value{
TRUE if the file was successfully untrashed, an error otherwise.
}
\description{
moves a file in Google Drive's "trash" folder, which is automatically
emptied after a set number of days, out and back into the user's Drive.
}

