% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upload.r
\name{upload_file}
\alias{upload_file}
\title{upload a file to Google Drive}
\usage{
upload_file(file_name, title = file_name, type = NULL,
  token = get_access_cred())
}
\arguments{
\item{file_name}{file name}

\item{title}{file title}

\item{type}{mime type of the file}

\item{token}{a valid OAuth2.0 token}
}
\value{
ID of the uploaded file
}
\description{
upload a file to Google Drive
}

