% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upload.r
\name{upload_file}
\alias{upload_file}
\title{upload a file to Google Drive}
\usage{
upload_file(file_name, title = basename(file_name), file_type = NULL,
  email = NULL, token = get_access_cred(), ...)
}
\arguments{
\item{file_name}{file name}

\item{title}{file title}

\item{file_type}{mime type of the file}

\item{email}{email address to add to permission list}

\item{token}{a valid OAuth2.0 token}

\item{...}{any other arguments to be passed to add_permission()}
}
\value{
ID of the uploaded file
}
\description{
upload a file to Google Drive
}

